# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GameForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/rules:
    get:
      operationId: listRules
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Rule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
    post:
      operationId: createRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleBody"
      responses:
        "201":
          description: Created rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /projects/{projectId}/rules/{ruleId}:
    patch:
      operationId: updateRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/RuleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
    delete:
      operationId: deleteRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/RuleId"
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/rules/ai-generate:
    post:
      operationId: aiGenerateRules
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerateBody"
      responses:
        "200":
          description: Generated rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rule"
  /projects/{projectId}/rules/{ruleId}/enhance:
    post:
      operationId: aiEnhanceRule
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Enhanced rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rule"
  /projects/{projectId}/rules/conflict-check:
    post:
      operationId: conflictCheckRules
      tags: [rules]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Conflict report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictReport"
...
  /projects/{projectId}/simulator/run:
    post:
      operationId: runSimulator
      tags: [simulator]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulatorRunBody"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulatorResult"
  /projects/{projectId}/simulator/playthrough:
    post:
      operationId: simulatorPlaythrough
      tags: [simulator]
      summary: Stream an AI playthrough narrative via SSE
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaythroughBody"
      responses:
        "200":
          description: SSE stream of narrative chunks
          content:
            text/event-stream: {}
...
  /projects/{projectId}/storyboard:
    get:
      operationId: listStoryboardNodes
      tags: [storyboard]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Storyboard nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoryboardNode"
    post:
      operationId: createStoryboardNode
      tags: [storyboard]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoryboardNodeBody"
      responses:
        "201":
          description: Created node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryboardNode"
  /projects/{projectId}/storyboard/{nodeId}:
    patch:
      operationId: updateStoryboardNode
      tags: [storyboard]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/NodeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoryboardNodeBody"
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryboardNode"
    delete:
      operationId: deleteStoryboardNode
      tags: [storyboard]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/NodeId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
    RuleId:
      name: ruleId
      in: path
      required: true
      schema:
        type: integer
...
    NodeId:
      name: nodeId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Rule:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: [string, "null"]
        priority:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, title, content, priority, createdAt, updatedAt]
    CreateRuleBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        priority:
          type: integer
      required: [title, content]
    UpdateRuleBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        priority:
          type: integer
    ConflictReport:
      type: object
      properties:
        conflicts:
          type: array
          items:
            type: object
            properties:
              ruleIds:
                type: array
                items:
                  type: integer
              severity:
                type: string
              description:
                type: string
              suggestion:
                type: string
            required: [ruleIds, severity, description]
        summary:
          type: string
      required: [conflicts, summary]
...
    AiGenerateBody:
      type: object
      properties:
        prompt:
          type: string
        count:
          type: integer
      required: [prompt]
...
    SimulatorRunBody:
      type: object
      properties:
        iterations:
          type: integer
        turns:
          type: integer
        startingResources:
          type: number
        incomePerTurn:
          type: number
        upkeepPerTurn:
          type: number
        randomness:
          type: number
      required: [iterations, turns]
    SimulatorResult:
      type: object
      properties:
        healthScore:
          type: integer
        verdict:
          type: string
        meanFinal:
          type: number
        bankruptRate:
          type: number
        turns:
          type: array
          items:
            type: object
            properties:
              turn:
                type: integer
              p10:
                type: number
              p50:
                type: number
              p90:
                type: number
            required: [turn, p10, p50, p90]
      required: [healthScore, verdict, meanFinal, bankruptRate, turns]
    PlaythroughBody:
      type: object
      properties:
        focus:
          type: string
...
    StoryboardNode:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        parentId:
          type: [integer, "null"]
        title:
          type: string
        content:
          type: [string, "null"]
        nodeType:
          type: string
        status:
          type: string
        color:
          type: [string, "null"]
        positionX:
          type: number
        positionY:
          type: number
        linkedRuleId:
          type: [integer, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, projectId, title, nodeType, status, positionX, positionY, createdAt, updatedAt]
    CreateStoryboardNodeBody:
      type: object
      properties:
        parentId:
          type: integer
        title:
          type: string
        content:
          type: string
        nodeType:
          type: string
        status:
          type: string
        color:
          type: string
        positionX:
          type: number
        positionY:
          type: number
        linkedRuleId:
          type: integer
      required: [title]
    UpdateStoryboardNodeBody:
      type: object
      properties:
        parentId:
          type: integer
        title:
          type: string
        content:
          type: string
        nodeType:
          type: string
        status:
          type: string
        color:
          type: string
        positionX:
          type: number
        positionY:
          type: number
        linkedRuleId:
          type: integer
...